Update Table Structure
Added the Primary keys. Moved the WhenCreated the the UserInfo
(accounts) table adn added the WhenInitialied and When Terminated to the
LoginTrack (sessions) table. Moved the Lockout status to the UserInfo
table. Added an AttemptStatus to the LoginAttempts to keep track of if
the attempt was successful or not. Added a unique constraint to username
in user info since it's not a primary key and not automatically
constrained to be unique.  Feel free to rename anything you need.
Cheers.