Restructured, Added Procedure
Changed the structure and names of the tables and columns. Used a stored
procedure rather than a trigger to accomplish the lockout required in
the assignment. Saved .ddl files for: creating the database, populating
it, and finally removing it all.

Also uploaded it to TEAM3_TT, along with test data.