Took parts of login procedure out into 2 triggers.
In addition to changing the dmd, I also saved the new design in the
Login_Authenticator.ddl file, and updated the .ddl file for removing
everything. Removed the .ddl file for populating test data. Commented
everything. (If you generate a new .dll for any reason, make sure to
check the "comments" box.)